Move page elements into a folder
