+++ src/ims/element/admin/acl/_admin_acl.py
@@ -20,7 +20,7 @@ Admin access control page.
 
 from ims.config import Configuration
 
-from ..._page import Page
+from ...page import Page
 
 
 __all__ = ()
