+++ src/ims/element/admin/types/_admin_types.py
@@ -20,7 +20,7 @@ Admin incident types page.
 
 from ims.config import Configuration
 
-from ..._page import Page
+from ...page import Page
 
 
 __all__ = ()
