+++ src/ims/element/incident/incident_template/_incident_template.py
@@ -20,7 +20,7 @@ Incident template page.
 
 from ims.config import Configuration
 
-from ..._page import Page
+from ...page import Page
 
 
 __all__ = ()
