+++ src/ims/element/incident/queue/_queue.py
@@ -28,7 +28,7 @@ from ims.ext.klein import KleinRenderable
 from ims.model import Event
 
 from ..queue_template._queue_template import title
-from ..._page import Page
+from ...page import Page
 
 
 __all__ = ()
