+++ src/ims/element/incident/queue_template/_queue_template.py
@@ -20,7 +20,7 @@ Dispatch queue template page.
 
 from ims.config import Configuration
 
-from ..._page import Page
+from ...page import Page
 
 
 __all__ = ()
