+++ src/ims/element/incident/report_template/_report_template.py
@@ -20,7 +20,7 @@ Incident report template page.
 
 from ims.config import Configuration
 
-from ..._page import Page
+from ...page import Page
 
 
 __all__ = ()
