+++ src/ims/element/incident/reports_template/_reports_template.py
@@ -20,7 +20,7 @@ Incident reports template page element.
 
 from ims.config import Configuration
 
-from ..._page import Page
+from ...page import Page
 
 
 __all__ = ()
