+++ src/ims/element/page/_page.py
@@ -29,10 +29,10 @@ from twisted.web.template import Tag, renderer, tags
 from ims.config import Configuration
 from ims.ext.klein import KleinRenderable
 
-from ._element import Element
 from .footer import FooterElement
 from .header import HeaderElement
 from .nav import NavElement
+from .._element import Element
 
 
 __all__ = ()
