+++ src/ims/element/root/_root.py
@@ -20,7 +20,7 @@ Server root page.
 
 from ims.config import Configuration
 
-from .._page import Page
+from ..page import Page
 
 
 __all__ = ()
